Syringe Holder for Light Imaging
This device allows you to mount a sample into a capillary (pump it in an agarose mix) then mounting the capillary in the dedicated plug insert and mounting
in the motor insert and then push the sample out (the capillary works like a syringe with a plunger inside the glass tube) to image the sample. 

There are four different inserts you can use with this deice, which allow you to image different samples. The four inserts will be interchangeable with the platform. 

The goal of this post is to properly show how to mount the platform and install the device properly.